more bug fixes
using Test1-20181021 to verify as do not yet have access to the one
Daniel used to verify. Some tests fail due to invalid header
configurations. More concerning is that I get segmentation faults every
now and again, and if I rerun the tests I get them on different tests.
The seg faults seem to happen at the end of the predictor, maybe
something to do with the freeing of the numpy arrays...

+++ ccsds123_i2_hlm/_predictor/predictorpy.cpp
@@ -6,7 +6,7 @@ namespace py = pybind11;
 PYBIND11_MODULE(_predictor, m, py::mod_gil_not_used())
 {
   py::class_<Predictor>(m, "Predictor")
-      .def(py::init<py::object, py::object, NumpyArr<int>, bool>())
+      .def(py::init<py::object, py::object, NumpyArr<long long>, bool>())
       .def("compress", &Predictor::compress)
       .def("save_data", &Predictor::save_data);
 }
